feat(landing): finalize landing page
