+++ app/components/avatar/index.tsx
@@ -1,5 +1,5 @@
 'use client';
-
+import React from 'react';
 import Image from 'next/image';
 
 interface AvatarProps {
