+++ app/components/index.tsx
@@ -2,5 +2,4 @@ export { default as Hero } from './hero';
 export { default as Avatar } from './avatar';
 export { default as Nav } from './navbar/Nav';
 export { default as Logo } from './navbar/Logo';
-export { default as Container } from './container';
 export { default as ThemeToggler } from './theme/ThemeToggler';
