+++ app/components/navbar/Logo.tsx
@@ -1,4 +1,4 @@
-'use client'
+'use client';
 import { useContext } from 'react';
 
 import { ThemeContext } from '@/app/providers';
@@ -10,17 +10,17 @@ import { useRouter } from 'next/navigation';
 const Logo = () => {
   const router = useRouter();
   // @ts-ignore
-  const { theme } = useContext(ThemeContext)
-  const isDark = theme === 'dark'
-  const logoImage = isDark ? '/images/logoDark.png' : '/images/logoLight.png'
+  const { theme } = useContext(ThemeContext);
+  const isDark = theme === 'dark';
+  const logoImage = isDark ? '/images/logoDark.png' : '/images/logoLight.png';
 
   return (
     <Image
       onClick={() => router.push('/')}
-      className="hidden md:block cursor-pointer px-4"
+      className="hidden md:block cursor-pointer"
       src={logoImage}
-      height="100"
-      width="100"
+      height="70"
+      width="70"
       alt="Logo"
     />
   );
