+++ app/components/navbar/Menu.tsx
@@ -1,0 +1,10 @@
+const NavBarMenu = () =>
+  <div className="navbar-center hidden lg:flex">
+    <ul className="menu menu-horizontal px-1">
+      <li><a>Sign Up</a></li>
+      <li><a>Login</a></li>
+    </ul>
+  </div>
+    ;
+
+export default NavBarMenu;
