+++ app/components/navbar/MenuItem.tsx
@@ -1,4 +1,5 @@
 'use client';
+import React from 'react';
 
 interface MenuItemProps {
   onClick: () => void;
