+++ app/components/navbar/Text.tsx
@@ -1,0 +1,5 @@
+import Link from 'next/link';
+
+const NavBarText = () => <Link href="/" className="px-4">Network Canvas</Link>;
+
+export default NavBarText;
