--- app/components/navbar/UserMenu.tsx
@@ -1,123 +1,0 @@
-'use client';
-
-import { signOut } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
-import { useCallback, useState } from 'react';
-import { AiOutlineMenu } from 'react-icons/ai';
-
-import Avatar from '@/app/components/avatar';
-import MenuItem from '@/app/components/navbar/MenuItem';
-
-interface UserMenuProps {
-}
-
-const UserMenu: React.FC<UserMenuProps> = () => {
-  const router = useRouter();
-  return (
-    <div className="relative">
-      <div className="flex flex-row items-center gap-3">
-        <div
-          onClick={() => console.log('Clicked')}
-          className="
-            hidden
-            md:block
-            text-sm
-            font-semibold
-            py-3
-            px-4
-            rounded-full
-            hover:bg-neutral-100
-            transition
-            cursor-pointer
-          "
-        >
-          Airbnb your home
-        </div>
-        <div
-          onClick={() => console.log('Open')}
-          className="
-          p-4
-          md:py-1
-          md:px-2
-          border-[1px]
-          border-neutral-200
-          flex
-          flex-row
-          items-center
-          gap-3
-          rounded-full
-          cursor-pointer
-          hover:shadow-md
-          transition
-          "
-        >
-          <AiOutlineMenu />
-          <div className="hidden md:block">
-            <Avatar src="" />
-          </div>
-        </div>
-      </div>
-      {true &&
-        <div
-          className="
-            absolute
-            rounded-xl
-            shadow-md
-            w-[40vw]
-            md:w-3/4
-            bg-white
-            overflow-hidden
-            right-0
-            top-12
-            text-sm
-          "
-        >
-          <div className="flex flex-col cursor-pointer">
-            {true ?
-              <>
-                <MenuItem
-                  label="My trips"
-                  onClick={() => router.push('/trips')}
-                />
-                <MenuItem
-                  label="My favorites"
-                  onClick={() => router.push('/favorites')}
-                />
-                <MenuItem
-                  label="My reservations"
-                  onClick={() => router.push('/reservations')}
-                />
-                <MenuItem
-                  label="My properties"
-                  onClick={() => router.push('/properties')}
-                />
-                <MenuItem
-                  label="Airbnb your home"
-                  onClick={() => {}}
-                />
-                <hr />
-                <MenuItem
-                  label="Logout"
-                  onClick={() => signOut()}
-                />
-              </>
-              :
-              <>
-                <MenuItem
-                  label="Login"
-                  onClick={() => {}}
-                />
-                <MenuItem
-                  label="Sign up"
-                  onClick={() => {}}
-                />
-              </>
-            }
-          </div>
-        </div>
-      }
-    </div>
-  );
-};
-
-export default UserMenu;
